LAB02/Lab1/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.1
matplotlib==3.8.2
numpy==1.26.2
